feat: 고객과 관련된 일을 처리하는 CustomerManager 역할 생성 (#2)
